docstringとREADMEの修正
